+++ test/cbmc/proofs/CheckOptionsOuter/Makefile.json
@@ -3,7 +3,7 @@
   "CBMCFLAGS":
   [
       "--unwind 1",
-      "--unwindset __CPROVER_file_local_FreeRTOS_TCP_IP_c_prvSingleStepTCPHeaderOptions.2:32"
+      "--unwindset __CPROVER_file_local_FreeRTOS_TCP_Reception_c_prvSingleStepTCPHeaderOptions.2:32"
   ],
   "OPT":
   [
@@ -12,7 +12,7 @@
   "OBJS":
   [
     "$(ENTRY)_harness.goto",
-    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_IP.goto",
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_Reception.goto",
     "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_WIN.goto",
     "$(FREERTOS_PLUS_TCP)/FreeRTOS_Stream_Buffer.goto",
     "$(FREERTOS_PLUS_TCP)/test/FreeRTOS-Kernel/tasks.goto",
