+++ test/cbmc/proofs/parsing/ProcessReceivedTCPPacket/Makefile.json
@@ -12,7 +12,11 @@
     "$(FREERTOS_PLUS_TCP)/FreeRTOS_IP.goto",
     "$(FREERTOS_PLUS_TCP)/FreeRTOS_IP_Utils.goto",
     "$(FREERTOS_PLUS_TCP)/FreeRTOS_IP_Timers.goto",
-    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_IP.goto"
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_IP.goto",
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_Transmission.goto",
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_State_Handling.goto",
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_Utils.goto",
+    "$(FREERTOS_PLUS_TCP)/FreeRTOS_TCP_Reception.goto"
   ],
   "INSTFLAGS":
   [
