+++ test/unit-test/TCPFilePaths.cmake
@@ -18,6 +18,10 @@ set( TCP_SOURCES
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_Sockets.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_Stream_Buffer.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_IP.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_Transmission.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_Reception.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_State_Handling.c"
+     "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_Utils.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_TCP_WIN.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_Tiny_TCP.c"
      "${CMAKE_CURRENT_LIST_DIR}/../../FreeRTOS_UDP_IP.c" )
